Add files via upload
The Company Information Finder is a Python program that allows users to search for information about a company on Google.com. The program takes the name of the company as input from the user, searches for it on Google.com, and outputs information about the company. The program uses the requests library to send a GET request to Google.com with the company name entered by the user as the search query. It then uses the BeautifulSoup library to parse the HTML content of the search results page and extract the relevant information about the company. The information is stored in a list of div elements with class BNeawe s3v9rd AP7Wnd. If the search results contain information about the company, the program prints it to the console. Otherwise, it prints a message indicating that no information was found.